#9000 SQLToolTaskVersionValidator has moved in model.sql
